refactor(register): register form
