+++ app/register/_components/RegisterForm.tsx
@@ -1,0 +1,89 @@
+"use client";
+
+import React from "react";
+
+import { Button } from "@/components/ui/button";
+import {
+  Form,
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+
+import { z } from "zod";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import { AuthService } from "@/services";
+
+const registerSchema = z
+  .object({
+    username: z.string().min(2).max(50),
+    mail: z.string().email(),
+  })
+  .refine((data) => data.username[0] !== "_", {
+    message: "username should not start with '_'",
+    path: ["username"],
+  });
+
+type RegisterSchema = z.infer<typeof registerSchema>;
+
+const RegisterForm = () => {
+  const form = useForm<RegisterSchema>({
+    resolver: zodResolver(registerSchema),
+  });
+
+  const onSubmit = async (data: RegisterSchema) => {
+    // Do something with the form values.
+    // ✅ This will be type-safe and validated.
+    console.log(data);
+
+    const { mail } = data;
+
+    const res = await AuthService.register({ mail });
+    console.log(res);
+  };
+
+  return (
+    <Form {...form}>
+      <form onSubmit={form.handleSubmit(onSubmit)} className="w-2/3 space-y-6">
+        <FormField
+          control={form.control}
+          name="username"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Username</FormLabel>
+              <FormControl>
+                <Input placeholder="username" {...field} />
+              </FormControl>
+              <FormDescription>
+                This is your public display name.
+              </FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <FormField
+          control={form.control}
+          name="mail"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Email</FormLabel>
+              <FormControl>
+                <Input placeholder="email" {...field} />
+              </FormControl>
+              <FormDescription>This is your email.</FormDescription>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+        <Button type="submit">Submit</Button>
+      </form>
+    </Form>
+  );
+};
+
+export default RegisterForm;
