+++ app/register/_components/index.ts
@@ -1,0 +1,4 @@
+import RegisterForm from "./RegisterForm";
+import OTPForm from "./OTPForm";
+
+export { RegisterForm, OTPForm };
